+++ appinfo/app.php
@@ -36,11 +36,4 @@
 
 namespace OCA\Onlyoffice\AppInfo;
 
-use OCP\App;
-
-/**
- * Administration settings
- */
-App::registerAdmin("onlyoffice", "settings");
-
 $app = \OC::$server->query(Application::class);